🗂Setup Base Multiplier in the Room database
- 🔌 Plug the value into the UI
- ♻️ Refactored to use Double as the BigDecimal is only useful for the string presentation, which is handled in the DoubleExtensions file now

+++ app/src/main/java/com/alejandrohcruz/currency/ui/base/listeners/RecyclerItemListener.kt
@@ -7,9 +7,9 @@ interface RecyclerItemListener {
     fun onItemSelected(
         currency: CurrencyEnum,
         position: Int,
-        newBaseMultiplier: BigDecimal?
+        newBaseMultiplier: Double?
     )
     fun onTextBeingEdited(position: Int)
     fun onTextNotBeingEdited(position: Int)
-    fun onBaseMultiplierChanged(newBaseMultiplier: BigDecimal)
+    fun onBaseMultiplierChanged(newBaseMultiplier: Double)
 }
