+++ app/src/main/java/com/alejandrohcruz/currency/utils/DoubleExtensions.kt
@@ -1,13 +1,12 @@
 package com.alejandrohcruz.currency.utils
 
-import java.math.BigDecimal
 import java.math.RoundingMode
 
 /**
  * @brief Trims out the decimal values for natural numbers and 00 for decimal numbers
  */
-fun BigDecimal.toPresentableString(): String {
-    return this.setScale(2, RoundingMode.HALF_UP) // round
+fun Double.toPresentableString(): String {
+    return this.toBigDecimal().setScale(2, RoundingMode.HALF_UP) // round
         .toPlainString()
         .removeSuffix(".00")
         .removeSuffix(".0")
